+++ terraform/modules/networking/variables.tf
@@ -1,0 +1,14 @@
+variable "location" {
+  description = "Azure region"
+  type        = string
+}
+
+variable "resource_group_name" {
+  description = "Resource Group name"
+  type        = string
+}
+
+variable "tags" {
+  description = "Common Azure tags"
+  type        = map(string)
+}